safedestroy can delete content hosts

+++ app/services/foreman_expire_hosts/safe_destroy.rb
@@ -8,7 +8,15 @@ module ForemanExpireHosts
     end
 
     def destroy!
-      subject.destroy!
+      # If Katello is installed, we can't just destroy the host
+      # but have to ask ForemanTasks to do this for us.
+      # See https://community.theforeman.org/t/how-to-properly-destroy-a-content-host/8621
+      # for reasoning.
+      if with_katello?
+        ForemanTasks.sync_task(::Actions::Katello::Host::Destroy, subject)
+      else
+        subject.destroy!
+      end
     rescue ActiveRecord::RecordNotDestroyed => invalid
       message = _('Failed to delete %{class_name} %{subject}: %{message} - Errors: %{errors}') % {
         :class_name => subject.class.name,
@@ -17,6 +25,15 @@ module ForemanExpireHosts
         :errors => invalid.record.errors.full_messages.to_sentence
       }
       Foreman::Logging.exception(message, invalid)
+      false
+    end
+
+    private
+
+    def with_katello?
+      Katello # rubocop:disable Lint/Void
+      true
+    rescue StandardError
       false
     end
   end
